[Feat] Network 모듈 Layer 분리 후 테스트 재적용
