+++ Projects/Core/Network/Testing/Sources/NetworkTesting.swift
@@ -1,0 +1,13 @@
+//
+//  NetworkTesting.swift
+//  CoreNetwork
+//
+//  Created by Greem on 6/29/25.
+//
+
+import CoreNetworkInterface
+import CoreNetwork
+
+public struct TempTestResponse: Codable {
+    public let id: Int
+}
